Move Ordinal to own file
